modularize triangular update.py
