+++ configs/train-af3-compare.yaml
@@ -8,9 +8,10 @@ model:
 train:
   _yaml_: "train/structure_only.yaml"
 
-  name: af3-edm-512
+  name: kfold-af3-edm
   out_dir: ./experiments/
   seed: 42
+  synchronize_seed: false # should be false when synchronize_sigmas=true
 
   # pytorch lightning trainer hparams
   trainer:
@@ -20,18 +21,18 @@ train:
     precision: bf16-mixed
     max_epochs: 100
     limit_train_batches: 1000
-    accumulate_grad_batches: 32
+    accumulate_grad_batches: 4
 
   training:
-    diffusion_batch_size: 32
+    diffusion_batch_size: 16
 
   data:
-    train_batch_size: 2
+    train_batch_size: 4
     max_tokens: 512
     lmdb_path: /cache/wykim_lab/kfold_data/kfold_rcsb_processed_v251120.lmdb/
     manifest_path: /cache/wykim_lab/kfold_data/manifests/all_manifest.pkl
     split_path: ./assets/splits/boltz1
-    num_workers: 4
+    num_workers: 8
 
     featurization_args:
       synchronize_ref_pos_augmentation: true
