--- src/kfold/model/layers/alphafold3/triangular_update/__init__.py
@@ -1,8 +1,0 @@
-from .attention import (
-    TriangleAttentionEndingNode,
-    TriangleAttentionStartingNode,
-)
-from .multiplication import (
-    TriangleMultiplicationIncoming,
-    TriangleMultiplicationOutgoing,
-)
