+++ src/kfold/model/layers/primitives/__init__.py
@@ -2,6 +2,14 @@ from .activation import SwiGLU
 from .attention import attention
 from .linear import Linear, LinearNoBias
 from .normalization import AdaLN, LayerNorm
+from .triangular_attention import (
+    TriangleAttentionEndingNode,
+    TriangleAttentionStartingNode,
+)
+from .triangular_multiplication import (
+    TriangleMultiplicationIncoming,
+    TriangleMultiplicationOutgoing,
+)
 
 __all__ = [
     "SwiGLU",
@@ -10,4 +18,8 @@ __all__ = [
     "AdaLN",
     "LayerNorm",
     "attention",
+    "TriangleAttentionStartingNode",
+    "TriangleAttentionEndingNode",
+    "TriangleMultiplicationOutgoing",
+    "TriangleMultiplicationIncoming",
 ]
